+++ src/integrations/virtualRoomIntegration/SelectVirtualRoomWidget.test.tsx
@@ -36,20 +36,24 @@ const scheduledMeetingMod: MeetingBe = createMockMeeting({
 
 describe('SelectVirtualRoomWidget', () => {
 	test('Should render properly - user has virtual rooms', async () => {
-		const store = useStore.getState();
-		store.setLoginInfo(sessionUser.id, sessionUser.name);
-		store.setCapabilities(createMockCapabilityList());
-		store.setRooms([temporaryRoomMod]);
-		store.setMeetings([scheduledMeetingMod]);
-		setup(
-			<SelectVirtualRoomWidgetComponent
-				onChange={jest.fn()}
-				defaultValue={{
-					label: temporaryRoomMod.name ?? '',
-					link: 'https://localhost/carbonio/focus-mode/meetings/scheduled-meeting-mod-test'
-				}}
-			/>
-		);
+		act(() => {
+			const store = useStore.getState();
+			store.setLoginInfo(sessionUser.id, sessionUser.name);
+			store.setCapabilities(createMockCapabilityList());
+			store.setRooms([temporaryRoomMod]);
+			store.setMeetings([scheduledMeetingMod]);
+		});
+		await act(async () => {
+			setup(
+				<SelectVirtualRoomWidgetComponent
+					onChange={jest.fn()}
+					defaultValue={{
+						label: temporaryRoomMod.name ?? '',
+						link: 'https://localhost/carbonio/focus-mode/meetings/scheduled-meeting-mod-test'
+					}}
+				/>
+			);
+		});
 
 		const selectVirtualRoom = screen.getByTestId('select_virtual_room');
 		expect(selectVirtualRoom).toBeInTheDocument();
@@ -62,9 +66,7 @@ describe('SelectVirtualRoomWidget', () => {
 		const store = useStore.getState();
 		store.setLoginInfo(sessionUser.id, sessionUser.name);
 		store.setCapabilities(createMockCapabilityList());
-		await act(async () => {
-			setup(<SelectVirtualRoomWidgetComponent onChange={() => null} defaultValue={undefined} />);
-		});
+		setup(<SelectVirtualRoomWidgetComponent onChange={() => null} defaultValue={undefined} />);
 
 		const noVirtualRoom = screen.getByTestId('no_virtual_room');
 		expect(noVirtualRoom).toBeInTheDocument();