+++ src/network/apis/UsersApi.test.ts
@@ -9,11 +9,9 @@ import * as uuid from 'uuid';
 import usersApi from './UsersApi';
 import useStore from '../../store/Store';
 import { createMockUser } from '../../tests/createMock';
-import { fetchResponse } from '../../tests/mocks/global';
+import { spyOnFetch } from '../../tests/jest-env-setup';
+import { RequestType } from '../../types/network/apis/IBaseAPI';
 import { UserBe } from '../../types/network/models/userBeTypes';
-
-const contentType = 'Content-Type';
-const applicationJson = 'application/json';
 
 const user: UserBe = createMockUser({ id: uuid.v6() });
 const user2: UserBe = createMockUser({ id: uuid.v6() });
@@ -21,20 +19,10 @@ const user2: UserBe = createMockUser({ id: uuid.v6() });
 describe('Users API', () => {
 	test('getUser is called correctly', async () => {
 		// Send getUser request
-		fetchResponse.mockResolvedValueOnce(user);
+		spyOnFetch.mockResolvedValueOnce(user);
 		await usersApi.getUser(user.id);
 
-		// Set appropriate headers
-		const headers = new Headers();
-		headers.append(contentType, applicationJson);
-
-		// Check if fetch is called with the correct parameters
-		expect(global.fetch).toHaveBeenCalledWith(`/services/chats/users/${user.id}`, {
-			method: 'GET',
-			headers,
-			body: undefined
-		});
-
+		expect(spyOnFetch).toHaveBeenCalledWith(`users/${user.id}`, RequestType.GET);
 		// Check if store is correctly updated
 		const store = useStore.getState();
 		expect(store.users[user.id]).toEqual(user);
@@ -42,21 +30,12 @@ describe('Users API', () => {
 
 	test('getUsers is called correctly', async () => {
 		// Send getUser request
-		fetchResponse.mockResolvedValueOnce([user, user2]);
+		spyOnFetch.mockResolvedValueOnce([user, user2]);
 		await usersApi.getUsers([user.id, user2.id]);
 
-		// Set appropriate headers
-		const headers = new Headers();
-		headers.append(contentType, applicationJson);
-
-		// Check if fetch is called with the correct parameters
-		expect(global.fetch).toHaveBeenCalledWith(
-			`/services/chats/users?userIds=${user.id}&userIds=${user2.id}`,
-			{
-				method: 'GET',
-				headers,
-				body: undefined
-			}
+		expect(spyOnFetch).toHaveBeenCalledWith(
+			`users?userIds=${user.id}&userIds=${user2.id}`,
+			RequestType.GET
 		);
 
 		// Check if store is correctly updated