+++ src/network/apis/UsersApi.ts
@@ -6,13 +6,13 @@
 
 import { forEach, join, map } from 'lodash';
 
-import BaseAPI from './BaseAPI';
 import useStore from '../../store/Store';
 import { RequestType } from '../../types/network/apis/IBaseAPI';
 import IUsersApi from '../../types/network/apis/IUsersApi';
 import { GetUserResponse, GetUsersResponse } from '../../types/network/responses/usersResponses';
+import { fetchAPI } from '../../utils/FetchUtils';
 
-class UsersApi extends BaseAPI implements IUsersApi {
+class UsersApi implements IUsersApi {
 	// Singleton design pattern
 	private static instance: IUsersApi;
 
@@ -25,7 +25,7 @@ class UsersApi extends BaseAPI implements IUsersApi {
 
 	public getUser(userId: string): Promise<GetUserResponse> {
 		const { setUserInfo } = useStore.getState();
-		return this.fetchAPI(`users/${userId}`, RequestType.GET).then((resp: GetUserResponse) => {
+		return fetchAPI(`users/${userId}`, RequestType.GET).then((resp: GetUserResponse) => {
 			setUserInfo(resp);
 			return resp;
 		});
@@ -34,12 +34,10 @@ class UsersApi extends BaseAPI implements IUsersApi {
 	public getUsers(userIds: string[]): Promise<GetUsersResponse> {
 		const { setUserInfo } = useStore.getState();
 		const ids = map(userIds, (id) => `userIds=${id}`);
-		return this.fetchAPI(`users?${join(ids, '&')}`, RequestType.GET).then(
-			(resp: GetUsersResponse) => {
-				forEach(resp, (user) => setUserInfo(user));
-				return resp;
-			}
-		);
+		return fetchAPI(`users?${join(ids, '&')}`, RequestType.GET).then((resp: GetUsersResponse) => {
+			forEach(resp, (user) => setUserInfo(user));
+			return resp;
+		});
 	}
 }