+++ src/settings/views/SettingsView.tsx
@@ -6,7 +6,7 @@
 
 import React, { ReactElement } from 'react';
 
-import { Spinner } from '@zextras/carbonio-shell-ui';
+import { Container, Spinner } from '@zextras/carbonio-design-system';
 
 import useStore from '../../store/Store';
 import Settings from '../components/Settings';
@@ -14,7 +14,11 @@ import Settings from '../components/Settings';
 const SettingsView = (): ReactElement => {
 	const id: string | undefined = useStore((store) => store.session.id);
 	if (!id) {
-		return <Spinner data-testid="spinner" />;
+		return (
+			<Container>
+				<Spinner color={'primary'} />
+			</Container>
+		);
 	}
 	return <Settings />;
 };